requirements.txt
python-dotenv
openai
requests
supabase
python-telegram-bot==20.0
